autogen-deep-research/pyproject.toml
[project]
name = "autogen-deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen>=0.9.10",
    "autogen-agentchat>=0.7.5",
    "autogen-ext[openai]>=0.7.5",
    "firecrawl-py>=4.3.6",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = ["jupyterlab>=4.4.9"]
